Add files via upload
Project files
The music data for the Music Genre Classification project can be found in the music_dataset_mod.csv file,
and the data legend is provided in the Music Data Legend.xlsx. 
To get started, you can download the Music Genre Classification with PCA - Project.ipynb notebook, open it in Jupyter Notebook, and run the first cell to import the required libraries.